bugfix for uploading files > 2GB
git-svn-id: https://libaws.svn.sourceforge.net/svnroot/libaws/trunk/libaws@162 e7c2904d-f150-0410-afc1-d11f7104a06f

+++ src/s3/s3object.h
@@ -34,7 +34,7 @@ public:
     typedef std::pair<std::string, std::list<std::string> > metadatapair_t;
 
     std::string      theContentType;
-    size_t           theContentLength;
+    uint64_t         theContentLength;
     metadata_t       theMetadata;
 
     // use either of the following memebers
